# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ArchiveTenantCustomRoleRequestDto": {
        "properties": {
          "expectedVersion": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "CreateTenantCustomRoleRequestDto": {
        "properties": {
          "description": {
            "maxLength": 500,
            "nullable": true,
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "permissionKeys": {
            "items": {
              "enum": [
                "tenant.membership.read",
                "tenant.membership.admin.invite",
                "tenant.membership.admin.suspend",
                "tenant.membership.admin.revoke",
                "tenant.membership.owner.promote",
                "tenant.membership.owner.demote",
                "tenant.security.session.read",
                "tenant.security.session.revoke",
                "tenant.rbac.permission.read",
                "tenant.rbac.role.read",
                "tenant.rbac.role.create",
                "tenant.rbac.role.update",
                "tenant.rbac.role.archive",
                "tenant.rbac.role.permission.grant",
                "tenant.rbac.role.permission.revoke",
                "tenant.rbac.assignment.read",
                "tenant.rbac.assignment.grant",
                "tenant.rbac.assignment.revoke"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "name",
          "permissionKeys"
        ],
        "type": "object"
      },
...
      "ReplaceTenantCustomRolePermissionsRequestDto": {
        "properties": {
          "expectedVersion": {
            "minimum": 1,
            "type": "number"
          },
          "permissionKeys": {
            "items": {
              "enum": [
                "tenant.membership.read",
                "tenant.membership.admin.invite",
                "tenant.membership.admin.suspend",
                "tenant.membership.admin.revoke",
                "tenant.membership.owner.promote",
                "tenant.membership.owner.demote",
                "tenant.security.session.read",
                "tenant.security.session.revoke",
                "tenant.rbac.permission.read",
                "tenant.rbac.role.read",
                "tenant.rbac.role.create",
                "tenant.rbac.role.update",
                "tenant.rbac.role.archive",
                "tenant.rbac.role.permission.grant",
                "tenant.rbac.role.permission.revoke",
                "tenant.rbac.assignment.read",
                "tenant.rbac.assignment.grant",
                "tenant.rbac.assignment.revoke"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": [
          "permissionKeys",
          "expectedVersion"
        ],
        "type": "object"
      },
...
      "TenantCustomRoleAssignmentResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "membershipId": {
            "format": "uuid",
            "type": "string"
          },
          "revokedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "roleId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "membershipId",
          "roleId",
          "createdAt"
        ],
        "type": "object"
      },
      "TenantCustomRoleResponseDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "normalizedName": {
            "type": "string"
          },
          "permissionKeys": {
            "items": {
              "enum": [
                "tenant.membership.read",
                "tenant.membership.admin.invite",
                "tenant.membership.admin.suspend",
                "tenant.membership.admin.revoke",
                "tenant.membership.owner.promote",
                "tenant.membership.owner.demote",
                "tenant.security.session.read",
                "tenant.security.session.revoke",
                "tenant.rbac.permission.read",
                "tenant.rbac.role.read",
                "tenant.rbac.role.create",
                "tenant.rbac.role.update",
                "tenant.rbac.role.archive",
                "tenant.rbac.role.permission.grant",
                "tenant.rbac.role.permission.revoke",
                "tenant.rbac.assignment.read",
                "tenant.rbac.assignment.grant",
                "tenant.rbac.assignment.revoke"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "version": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "normalizedName",
          "version",
          "permissionKeys"
        ],
        "type": "object"
      },
...
      "UpdateTenantCustomRoleRequestDto": {
        "properties": {
          "description": {
            "maxLength": 500,
            "nullable": true,
            "type": "string"
          },
          "expectedVersion": {
            "minimum": 1,
            "type": "number"
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "name",
          "expectedVersion"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Supported Booking OS HTTP API",
    "title": "Booking OS API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tenant/rbac/memberships/{membershipId}/roles": {
      "get": {
        "operationId": "listTenantMembershipRbacRoles",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      }
    },
    "/api/tenant/rbac/memberships/{membershipId}/roles/{roleId}": {
      "delete": {
        "operationId": "revokeTenantMembershipRbacRole",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      },
      "post": {
        "operationId": "grantTenantMembershipRbacRole",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCustomRoleAssignmentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      }
    },
...
    "/api/tenant/rbac/roles": {
      "get": {
        "operationId": "listTenantRbacRoles",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      },
      "post": {
        "operationId": "createTenantRbacRole",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantCustomRoleRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      }
    },
    "/api/tenant/rbac/roles/{roleId}": {
      "delete": {
        "operationId": "archiveTenantRbacRole",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveTenantCustomRoleRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      },
      "get": {
        "operationId": "getTenantRbacRole",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      },
      "patch": {
        "operationId": "updateTenantRbacRole",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantCustomRoleRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      }
    },
    "/api/tenant/rbac/roles/{roleId}/permissions": {
      "put": {
        "operationId": "replaceTenantRbacRolePermissions",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceTenantCustomRolePermissionsRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCustomRoleResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "tenant-rbac"
        ]
      }
    }
  },
...
}